+++ config.conf
@@ -25,5 +25,22 @@ http {
     
     server {
         # configuration of HTTP virtual server 2
+        listen        two.example.com;
+		server_name   two.example.com  www.two.example.com;
+
+		error_page  404  /404.html;
+
+		location  /404.html {
+            root  /www/errors;
+            autoindex    off;
+        }
+        location /one {
+            # configuration for processing URIs starting with '/one'
+			autoindex    on;
+        }
+        location /two {
+            root;
+            # configuration for processing URIs starting with '/two'
+        }
     }
 }
